# Remove inheritance from ContextDecorator