--- a/python/nvtx/nvtx.py
+++ b/python/nvtx/nvtx.py
@@ -5,6 +5,7 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 from contextlib import ContextDecorator
+from functools import wraps
 
 from nvtx._lib import (
     Domain,
@@ -17,7 +18,7 @@ from nvtx._lib import (
 )
 
 
-class annotate(ContextDecorator):
+class annotate:
     """
     Annotate code ranges using a context manager or a decorator.
     """
@@ -89,7 +90,15 @@ class annotate(ContextDecorator):
     def __call__(self, func):
         if not self.attributes.message:
             self.attributes.message = func.__name__
-        return super().__call__(func)
+
+        @wraps(func)
+        def inner(*args, **kwargs):
+            libnvtx_push_range(self.attributes, self.domain.handle)
+            result = func(*args, **kwargs)
+            libnvtx_pop_range(self.domain.handle)
+            return result
+
+        return inner
 
 
 def mark(message=None, color="blue", domain=None, category=None):