Don't inherit from ContextDecorator. This improves the performance of `annotate` when used as a decorator. It also leads to fewer functions in the profile:

Before:

```python
In [1]: from nvtx import annotate

In [2]: @annotate()
   ...: def  foo():
   ...:     pass
   ...:

In [3]: %timeit foo()
558 ns ± 0.207 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %prun foo()
         10 function calls in 0.000 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 contextlib.py:72(inner)
        1    0.000    0.000    0.000    0.000 nvtx.py:81(__enter__)
        1    0.000    0.000    0.000    0.000 <string>:1(<module>)
        1    0.000    0.000    0.000    0.000 nvtx.py:85(__exit__)
        1    0.000    0.000    0.000    0.000 {nvtx._lib.lib.push_range}
        1    0.000    0.000    0.000    0.000 contextlib.py:59(_recreate_cm)
        1    0.000    0.000    0.000    0.000 {nvtx._lib.lib.pop_range}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        1    0.000    0.000    0.000    0.000 <ipython-input-2-49e4e1689ec5>:1(foo)
```

After:

```python
In [1]: from nvtx import annotate

In [2]: @annotate()
   ...: def  foo():
   ...:     pass
   ...:

In [3]: %timeit foo()
306 ns ± 0.0755 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %prun foo()
         7 function calls in 0.000 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 nvtx.py:94(inner)
        1    0.000    0.000    0.000    0.000 <string>:1(<module>)
        1    0.000    0.000    0.000    0.000 {nvtx._lib.lib.push_range}
        1    0.000    0.000    0.000    0.000 <ipython-input-2-49e4e1689ec5>:1(foo)
        1    0.000    0.000    0.000    0.000 {nvtx._lib.lib.pop_range}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```